% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{do_least_squares_with_lsfit}
\alias{do_least_squares_with_lsfit}
\title{Get the OLS solution using \code{lsfit()}}
\usage{
do_least_squares_with_lsfit(y, X)
}
\arguments{
\item{y}{our response, which is an n-vector}

\item{X}{our data matrix, which is n by p}
}
\description{
Get the OLS solution using \code{lsfit()}
}