% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{say_hello}
\alias{say_hello}
\title{Say hello to someone}
\usage{
say_hello(name, exclamation = TRUE)
}
\arguments{
\item{name}{Name of a person}

\item{exclamation}{Whether to include an exclamation mark}
}
\description{
Say hello to someone
}