% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\docType{package}
\name{frombookdown-package}
\alias{frombookdown}
\alias{frombookdown-package}
\title{A Package That Does Least Squares}
\description{
This package does least squares.  But what's actually interesting about it is that it was defined as a \code{bookdown} using the \code{litr} package.
}
\seealso{
\code{do_least_squares}
}
\author{
\strong{Maintainer}: Jacob Bien \email{jbien@usc.edu}

}
\keyword{internal}