% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\docType{package}
\name{frombookdown-package}
\alias{frombookdown}
\alias{frombookdown-package}
\title{A Package Defined in a Bookdown}
\description{
This package uses \code{litr} to define an R package through writing a \code{bookdown}.
}
\author{
\strong{Maintainer}: First Last \email{you@gmail.com}

}
\keyword{internal}